slow dwt impl

+++ emu/plugin/cyc_model.h
@@ -1,7 +1,7 @@
 #pragma once
 #include <stdint.h>
 
-typedef uint16_t op_code_t;
+typedef uint32_t op_code_t;
 typedef uint32_t address_t;
 
 uint32_t get_instruction_cycles(op_code_t op_code);
